AP_Torqeedo: multi backend support
